Extract listLinkInfos
* no longer on reconcile struct
* replace nested if with guard clauses
